tests/requirements-test.txt
# tests/requirements-test.txt
# Testing dependencies

pytest>=7.0.0
pytest-cov>=4.0.0
pytest-timeout>=2.1.0
pytest-xdist>=3.0.0
numpy>=1.21.0
psutil>=5.9.0
pyyaml>=6.0
